fix(StubCovers): cleanup folder structure
