+++ Editor/CVRFury/StubCovers/VRC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 29924c0fe9c8d854da21c1b21218ce62
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
