+++ Editor/CVRFury/StubCovers/VRC/VRCAvatarDescriptorEditor.cs.meta
