+++ Editor/CVRFury/StubCovers/VRC/VRCContactSender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5dc203199c7621446a7be84d7eaaaf38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
