+++ Editor/CVRFury/StubCovers/VRCFury.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 789470c8f20fd2d44be543735bf41e11
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
