--- Editor/CVRFury/StubCovers/VRCFuryGeneralStubEditor.cs
@@ -1,74 +1,0 @@
-#if UNITY_EDITOR
-using System.Collections.Generic;
-using UnityEditor;
-using UnityEditor.UIElements;
-using UnityEngine;
-using UnityEngine.UIElements;
-using uk.novavoidhowl.dev.vrcstub;
-using Constants = uk.novavoidhowl.dev.cvrfury.packagecore.Constants;
-using static uk.novavoidhowl.dev.cvrfury.packagecore.CoreUtils;
-
-namespace uk.novavoidhowl.dev.cvrfury
-{
-  public abstract class VRCFuryGeneralStubEditor : Editor
-  {
-    public override VisualElement CreateInspectorGUI()
-    {
-      // Create the root VisualElement
-      var rootVisualElement = new VisualElement();
-
-      // set the class of the root element to allow styling
-      rootVisualElement.AddToClassList("vrcfury-stub-inspector");
-
-      // load base UXML
-      var baseTree = Resources.Load<VisualTreeAsset>(
-        Constants.PROGRAM_DISPLAY_NAME + "/StubCovers/UnityUXML/VRCFuryGeneralStub"
-      );
-
-      // Check if the UXML file was loaded
-      if (baseTree == null)
-      {
-        CoreLogError(
-          "Failed to load UXML file at 'UnityUXML/VRCFuryGeneralStub'. Please ensure the file exists at the specified path."
-        );
-        // If the UXML file was not loaded add a new label to the root.
-        rootVisualElement.Add(new Label("CRITICAL ERROR : UXML could not be loaded."));
-        return rootVisualElement;
-      }
-
-      // Load and apply the stylesheet
-      var stylesheet = Resources.Load<StyleSheet>(
-        Constants.PROGRAM_DISPLAY_NAME + "/StubCovers/UnityStyleSheets/VRCFuryGeneralStub"
-      );
-
-      // Check if the StyleSheet was loaded
-      if (stylesheet == null)
-      {
-        CoreLogError(
-          "Failed to load StyleSheet at 'UnityStyleSheets/VRCFuryGeneralStub'. Please ensure the file exists at the specified path."
-        );
-        // If the StyleSheet was not loaded add a new label to the root.
-        rootVisualElement.Add(new Label("CRITICAL ERROR : StyleSheet could not be loaded."));
-        return rootVisualElement;
-      }
-
-      // Instantiate the UXML tree
-      var ToolSetup = baseTree.Instantiate();
-
-      // Create a temporary list to hold the children
-      List<VisualElement> children = new List<VisualElement>(ToolSetup.Children());
-
-      // Add the children of the instantiated UXML to the root
-      foreach (var child in children)
-      {
-        rootVisualElement.Add(child);
-      }
-
-      // Apply the StyleSheet
-      rootVisualElement.styleSheets.Add(stylesheet);
-
-      return rootVisualElement;
-    }
-  }
-}
-#endif
