+++ package.json
@@ -22,5 +22,5 @@
   },
   "unity": "2021.3",
   "unityRelease": "23f1",
-  "version": "0.178.0-alpha"
+  "version": "0.178.1-alpha"
 }
